# Improve disk IO speeds, optimize FAT Implementation memory allocation & speeds