--- a/source/Cosmos.HAL2/BlockDevice/ATA_PIO.cs
+++ b/source/Cosmos.HAL2/BlockDevice/ATA_PIO.cs
@@ -405,7 +405,7 @@ namespace Cosmos.HAL.BlockDevice
 			SelectSector(aBlockNo, aBlockCount);
 			SendCmd(LBA48Bit ? Cmd.ReadPioExt : Cmd.ReadPio);
             IOPort.Read8(IO.Data, aData);
-		}
+        }
 
         /// <summary>
         /// Writes the specific block of data using the starting block,