@GoldenretriverYT 2023-12-10
Draft for now cause I will look into speeding up writes too