@GoldenretriverYT 2023-12-10
![image](https://github.com/CosmosOS/Cosmos/assets/49924528/15bfed4c-fe3f-40e9-a379-15638ece7366)
Even a 1mb file took 77s before (under my testing after patching the memory leaks)